% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-concrete.R
\docType{data}
\name{concrete}
\alias{concrete}
\title{A concrete dataset for the KTPmars package.}
\format{
A data frame with 100 rows and 9 variables:
\describe{
\item{ConcreteCompressiveStrength}{response variable}
\item{Cement}{explanatory variable}
\item{BlastFurnaceSlag}{explanatory variable}
\item{FlyAsh}{explanatory variable}
\item{Water}{explanatory variable}
\item{Superplasticizer}{explanatory variable}
\item{CoarseAggregate}{explanatory variable}
\item{FineAggregate}{explanatory variable}
\item{Age}{explanatory variable}
}
}
\usage{
concrete
}
\description{
A dataset of size N=100 with n=8 explanatory variables,
and a response variable that is Concrete compressive strength
}
\keyword{datasets}
