% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{fwd_stepwise}
\alias{fwd_stepwise}
\title{Title}
\usage{
fwd_stepwise(y, x, control = mars.control())
}
\arguments{
\item{y}{y}

\item{x}{x}

\item{control}{cc}
}
\value{
a list
}
\description{
Title
}
\examples{
fwd <- fwd_stepwise(y=marstestdata$y,x=marstestdata[,-1])

}
