% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{h}
\alias{h}
\title{Hinge function for predictor value adjustment.}
\usage{
h(x, s, t)
}
\arguments{
\item{x}{a vector of predictor value.}

\item{s}{a numeric value of either 1 or -1, indicating sign of hinge function.}

\item{t}{value of split point, the knot of hinge function.}
}
\value{
a vector of numeric values that adjust values of given predictor
vector according to the hinge function.
}
\description{
Hinge function for predictor value adjustment.
}
\examples{
h(x=concrete$Cement,s=1,t=300)

}
\references{
\enumerate{
\item 
}

Paper: Multivariate Adaptive Regression Splines

Author(s): Jerome H. Friedman

Source: The Annals of Statistics , Mar., 1991, Vol. 19, No. 1 (Mar., 1991), pp. 1-67

Published by: Institute of Mathematical Statistics Stable

URL: https://www.jstor.org/stable/2241837
\enumerate{
\item 
}

Github repo of STAT 360 by Becky Lin:

https://github.com/Becky07/STAT360
}
