% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{init_B}
\alias{init_B}
\title{Initialization function for basis function matrix in forward selection}
\usage{
init_B(N, Mmax)
}
\arguments{
\item{N}{number of observations.}

\item{Mmax}{Mmax from an object of class 'mars.control', the maximum basis
functions will be added in forward stepwise function.}
}
\value{
a data frame with first column of 1s and rest column as NA.
}
\description{
Initialization function for basis function matrix in forward selection
}
\examples{
Bmatrix <- init_B(100,10)

}
\references{
i)

Paper: Multivariate Adaptive Regression Splines

Author(s): Jerome H. Friedman

Source: The Annals of Statistics , Mar., 1991, Vol. 19, No. 1 (Mar., 1991), pp. 1-67

Published by: Institute of Mathematical Statistics Stable

URL: https://www.jstor.org/stable/2241837

ii)

Github repo of STAT 360 by Becky Lin:

https://github.com/Becky07/STAT360
}
