% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mars.R
\name{make_B}
\alias{make_B}
\title{Function which uses Bfuncs and data frame to reproduce basis function matrix}
\usage{
make_B(X, Bfuncs)
}
\arguments{
\item{X}{a model matrix (predictors only).}

\item{Bfuncs}{a list of Bfuncs information, usually extracts from output of
'mars' object.}
}
\value{
a basis function matrix.
}
\description{
Function which uses Bfuncs and data frame to reproduce basis function matrix
}
\examples{
    tmars <- mars(ConcreteCompressiveStrength~.,data=concrete,control =
    mars.control(10,3,FASLE))
    tr <- terms(ConcreteCompressiveStrength~.,data=concrete)
    tr <- delete.response(tr)
    mf <- model.frame(tr,concrete)
    mt <- attr(mf, "terms")
    tX <- model.matrix(mt, mf)[,-1]
    tB <- make_B(tX,tmars$Bfuncs)

}
\references{
i)

Paper: Multivariate Adaptive Regression Splines

Author(s): Jerome H. Friedman

Source: The Annals of Statistics , Mar., 1991, Vol. 19, No. 1 (Mar., 1991), pp. 1-67

Published by: Institute of Mathematical Statistics Stable

URL: https://www.jstor.org/stable/2241837

ii)

Github repo of STAT 360 by Becky Lin:

https://github.com/Becky07/STAT360
}
