% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{mars.control}
\alias{mars.control}
\title{Helper for 'mars.control' objects}
\usage{
mars.control(Mmax = 2, d = 3, trace = FALSE)
}
\arguments{
\item{Mmax}{Maximum number of basis functions. Should be an even integer.
Default value is 2.}

\item{d}{The parameter used in calculation of Generalized cross-validation.
Friedman suggests that d = 3 works well, thus default value is 3.}

\item{trace}{A True or False value that allows user to see the candidate subset
basis functions that might reduce LOF (RSS) with GCV criterion in backward
selection procedure.}
}
\value{
an object of class 'mars.control', that is a list of parameters
}
\description{
This function constructs a \code{mars.control} object that specifies
parameters used in the model fitting procedure.
}
\examples{
mc<-mars.control(Mmax=10,d=3,trace=FALSE)

}
\references{
i)

Paper: Multivariate Adaptive Regression Splines

Author(s): Jerome H. Friedman

Source: The Annals of Statistics , Mar., 1991, Vol. 19, No. 1 (Mar., 1991), pp. 1-67

Published by: Institute of Mathematical Statistics Stable

URL: https://www.jstor.org/stable/2241837

ii)

Github repo of STAT 360 by Becky Lin:

https://github.com/Becky07/STAT360
}
