% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mars.R
\name{plot.mars}
\alias{plot.mars}
\title{Plot implementatioN - S3Method for an object of class 'mars'}
\usage{
\method{plot}{mars}(object)
}
\arguments{
\item{object}{an object of class 'mars'}
}
\value{
produce 3 plots for the fitted model using basis functions:
Residuals vs. Fitted values plot, histogram of residuals and normal Q-Q plot
of residuals
}
\description{
To produce graphics of the fitted model.
}
\examples{
mm<-mars(ConcreteCompressiveStrength~.,data=concrete,control = mars.control(
Mmax=10,d=3,trace=FALSE))
plots <- plot(mm)

}
\references{
i)

Paper: Multivariate Adaptive Regression Splines

Author(s): Jerome H. Friedman

Source: The Annals of Statistics , Mar., 1991, Vol. 19, No. 1 (Mar., 1991), pp. 1-67

Published by: Institute of Mathematical Statistics Stable

URL: https://www.jstor.org/stable/2241837

ii)

Github repo of STAT 360 by Becky Lin:

https://github.com/Becky07/STAT360
}
