% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mars.R
\name{predict.mars}
\alias{predict.mars}
\title{Predict implementatioN - S3Method for an object of class 'mars'}
\usage{
\method{predict}{mars}(object, newdata)
}
\arguments{
\item{object}{an object of class 'mars'.}

\item{newdata}{a data frame with response and predictor variables.}
}
\value{
a vector of predicted response value.
}
\description{
Predict implementatioN - S3Method for an object of class 'mars'
}
\examples{
mm<-mars(ConcreteCompressiveStrength~.,data=concrete,control = mars.control(
Mmax=10,d=3,trace=FALSE))
predictedvalues <- predict(mm)

}
\references{
i)

Paper: Multivariate Adaptive Regression Splines

Author(s): Jerome H. Friedman

Source: The Annals of Statistics , Mar., 1991, Vol. 19, No. 1 (Mar., 1991), pp. 1-67

Published by: Institute of Mathematical Statistics Stable

URL: https://www.jstor.org/stable/2241837

ii)

Github repo of STAT 360 by Becky Lin:

https://github.com/Becky07/STAT360
}
