% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{split_points}
\alias{split_points}
\title{A function that picks unique, available split point value in a vector of
predictor value.}
\usage{
split_points(xvar, Bm)
}
\arguments{
\item{xvar}{a vector of predictor value.}

\item{Bm}{mth column in basis function matrix.}
}
\value{
a vector of sorted, unique and valid split points, excluding the
highest value.
}
\description{
A function that picks unique, available split point value in a vector of
predictor value.
}
\examples{
fwdtest <- fwd_stepwise(y=concrete$ConcreteCompressiveStrength,x=concrete[,-1],
control = mars.control(6,3,FALSE))

sp <- split_points(concrete$Cement,fwdtest$B[,3])

}
\references{
\enumerate{
\item 
}

Paper: Multivariate Adaptive Regression Splines

Author(s): Jerome H. Friedman

Source: The Annals of Statistics , Mar., 1991, Vol. 19, No. 1 (Mar., 1991), pp. 1-67

Published by: Institute of Mathematical Statistics Stable

URL: https://www.jstor.org/stable/2241837
\enumerate{
\item 
}

Github repo of STAT 360 by Becky Lin:

https://github.com/Becky07/STAT360
}
