% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mars.R
\name{summary.mars}
\alias{summary.mars}
\title{Title}
\usage{
\method{summary}{mars}(object)
}
\arguments{
\item{object}{a mars object}
}
\value{
lm summary plus information of basis function
}
\description{
Title
}
\examples{
summary<-summary(testmars)
}
