% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{validate_mars.control}
\alias{validate_mars.control}
\title{Validator for 'mars.control' objects}
\usage{
validate_mars.control(control)
}
\arguments{
\item{control}{a list of values Mmax, d and trace.}
}
\value{
the original list if Mmax, d and trace are valid for the list to be
assigned to an object of class 'mars.control'. Otherwise,
either returning errors or coercing the values while sending warnings.
}
\description{
This function validates a list object if it is able to become an object of
class 'mars.control'.
}
\examples{
validmc <- validate_mars.control(list(Mmax=10,d=3,trace=TRUE))
}
\references{
i)

Paper: Multivariate Adaptive Regression Splines

Author(s): Jerome H. Friedman

Source: The Annals of Statistics , Mar., 1991, Vol. 19, No. 1 (Mar., 1991), pp. 1-67

Published by: Institute of Mathematical Statistics Stable

URL: https://www.jstor.org/stable/2241837

ii)

Github repo of STAT 360 by Becky Lin:

https://github.com/Becky07/STAT360
}
