% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-yacht.R
\docType{data}
\name{yacht}
\alias{yacht}
\title{A yacht dataset for the KTPmars package.}
\format{
A data frame with 100 rows and 7 variables:
\describe{
\item{RRPUWOD}{response variable}
\item{LPOTCOB}{explanatory variable}
\item{PrismaticCoefficient}{explanatory variable}
\item{LengthDisplacementRatio}{explanatory variable}
\item{BeamDraughtRatio}{explanatory variable}
\item{LengthBeamRatio}{explanatory variable}
\item{FroudeNumber}{explanatory variable}
}
}
\usage{
yacht
}
\description{
A dataset of size N=100 with n=6 explanatory variables,
and a response variable that is Residuary resistance per unit weight of
displacement (RRPUWOD). Predictor LPOTCOB means Longitudinal position of the
center of buoyancy. All other 5 predictors have original names.
}
\keyword{datasets}
