% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.mars.R
\name{anova.mars}
\alias{anova.mars}
\title{anova implementation}
\usage{
\method{anova}{mars}(object)
}
\arguments{
\item{object}{a mars object}
}
\value{
anova information regarding the mars object
}
\description{
anova implementation
}
\examples{
mm<-mars(y~.,data=marstestdata)
anova(mm)
}
