% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova.mars.R
\name{anova.mars}
\alias{anova.mars}
\title{Anova implementation - S3Method for an object of class 'mars'}
\usage{
\method{anova}{mars}(object)
}
\arguments{
\item{object}{a mars object.}
}
\value{
anova() information regarding the mars object, similar to anova output
of an object of class 'lm', and information of Basis functions.
}
\description{
To provide useful information about the basis functions.
}
\examples{
mm<-mars(ConcreteCompressiveStrength~.,data=concrete,control = mars.control(
Mmax=10,d=3,trace=FALSE))
output <- anova(mm)

}
\references{
i)

Paper: Multivariate Adaptive Regression Splines

Author(s): Jerome H. Friedman

Source: The Annals of Statistics , Mar., 1991, Vol. 19, No. 1 (Mar., 1991), pp. 1-67

Published by: Institute of Mathematical Statistics Stable

URL: https://www.jstor.org/stable/2241837

ii)

Github repo of STAT 360 by Becky Lin:

https://github.com/Becky07/STAT360
}
