% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{fwd_stepwise}
\alias{fwd_stepwise}
\title{Forward stepwise function based on Algorithm 2 by Friedman}
\usage{
fwd_stepwise(y, x, control = mars.control())
}
\arguments{
\item{y}{the vector of response variable used.}

\item{x}{the model matrix used.}

\item{control}{an object of class 'mars.control': a list of Mmax, d, and trace,
see helper file mars.control() for more information.}
}
\value{
a list containing the following information:
\itemize{
\item{y:}{ the column vector of response variable.}
\item{B:}{ a matrix that contains information of basis functions determined by
forward selection. Each column of basis functions are
products of hinge functions with information of sign, number of predictor and
split point given in Bfuncs in order.}
\item{Bfuncs:}{ a list containing information on how each basis function
is created, each element of Bfuncs correspond to the hinge functions that are
used to create each column of B. Inside each element of Bfuncs, there could be
null, 1 or multiples rows. Each row represents information of each hinge function.}
}
}
\description{
Using a given maximum number of basis functions we wish to have except the
intercept, builds basis function matrix B by adding pair of new basis function.
Function will loop over all predictor variables and available split points and
determine the right pair basis functions that produce the least LOF (RSS) standard
with GCV criterion in each iteration.
}
\examples{
fwdtest <- fwd_stepwise(y=concrete$ConcreteCompressiveStrength,x=concrete[,-1],
control = mars.control(8,3,FALSE))

}
\references{
i)

Paper: Multivariate Adaptive Regression Splines

Author(s): Jerome H. Friedman

Source: The Annals of Statistics , Mar., 1991, Vol. 19, No. 1 (Mar., 1991), pp. 1-67

Published by: Institute of Mathematical Statistics Stable

URL: https://www.jstor.org/stable/2241837

ii)

Github repo of STAT 360 by Becky Lin:

https://github.com/Becky07/STAT360
}
