% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{LOF}
\alias{LOF}
\title{LOF function with GCV criterion that produces}
\usage{
LOF(form, data, control)
}
\arguments{
\item{form}{an object of class 'formula': Specifically, a linear regression
formula that has the response variable's name and indicates the predictors.
It is similar to the formula used in lm(), but must add -1 at the end to
avoid an extra intercept column.}

\item{data}{a data frame with response and predictor variables.}

\item{control}{an object of class 'mars.control': a list of Mmax, d, and trace,
see helper file mars.control() for more information.}
}
\value{
value of LOF with GCV criterion.
}
\description{
LOF function with GCV criterion that produces
}
\examples{
tlof <- LOF(form = ConcreteCompressiveStrength~.-1,data = concrete,
control=mars.control())

}
\references{
\enumerate{
\item 
}

Paper: Multivariate Adaptive Regression Splines

Author(s): Jerome H. Friedman

Source: The Annals of Statistics , Mar., 1991, Vol. 19, No. 1 (Mar., 1991), pp. 1-67

Published by: Institute of Mathematical Statistics Stable

URL: https://www.jstor.org/stable/2241837
\enumerate{
\item 
}

Github repo of STAT 360 by Becky Lin:

https://github.com/Becky07/STAT360
}
