% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mars.R
\name{new_mars.control}
\alias{new_mars.control}
\title{Constructor for 'mars.control' objects}
\usage{
new_mars.control(control)
}
\arguments{
\item{control}{a list of values Mmax, d and trace, see ?mars.control for more
information.}
}
\value{
an object of class 'mars.control'.
}
\description{
This function construct a 'mars.control' object that specifies
parameters used in the model fitting procedure for users familiar to mars.
}
\examples{
testnewmc <- new_mars.control(list(Mmax=as.integer(10),d=3,trace=TRUE))

}
\references{
i)

Paper: Multivariate Adaptive Regression Splines

Author(s): Jerome H. Friedman

Source: The Annals of Statistics , Mar., 1991, Vol. 19, No. 1 (Mar., 1991), pp. 1-67

Published by: Institute of Mathematical Statistics Stable

URL: https://www.jstor.org/stable/2241837

ii)

Github repo of STAT 360 by Becky Lin:

https://github.com/Becky07/STAT360
}
