% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mars.R
\name{print.mars}
\alias{print.mars}
\title{Print implementatioN - S3Method for an object of class 'mars'}
\usage{
\method{print}{mars}(object)
}
\arguments{
\item{object}{an object of class 'mars'.}
}
\value{
prints matched function call and coefficients, similar to the print()
output of an object of class 'lm', and information of basis functions.
}
\description{
Print implementatioN - S3Method for an object of class 'mars'
}
\examples{
mm<-mars(ConcreteCompressiveStrength~.,data=concrete,control = mars.control(
Mmax=10,d=3,trace=FALSE))
result <- print(mm)

}
\references{
i)

Paper: Multivariate Adaptive Regression Splines

Author(s): Jerome H. Friedman

Source: The Annals of Statistics , Mar., 1991, Vol. 19, No. 1 (Mar., 1991), pp. 1-67

Published by: Institute of Mathematical Statistics Stable

URL: https://www.jstor.org/stable/2241837

ii)

Github repo of STAT 360 by Becky Lin:

https://github.com/Becky07/STAT360
}
